+++ library/src/main/java/com/pengrad/telegrambot/request/BanChatMember.java
@@ -9,7 +9,7 @@ public class BanChatMember extends BaseRequest<BanChatMember, BaseResponse> {
         add("chat_id", chatId).add("user_id", userId);
     }
 
-    public BanChatMember untilDate(long untilDate) {
+    public BanChatMember untilDate(int untilDate) {
         return add("until_date", untilDate);
     }
